+++ CHANGELOG.md
@@ -22,6 +22,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   `chfl_cell_shape_t`, `chfl_cell_shape`, and `chfl_cell_set_shape`.
 * All the floating point data uses doubles instead of floats. This concerns
   atomic data, positions and velocities.
+* Add "Selection::string" function and the corresponding `chfl_selection_string`
+  to get the string used to build a selection.
 
 ### Changes in supported formats